fix(openapi): correct merged response schemas and diagnostics
Emit anyOf for merged duplicate response bodies and warn when canonical
responses reuse the same name.

Preserve precise Zod lazy conversion warnings while keeping root fallback
diagnostics consistent.
